all floats are now doubles
note, this is bad for speed and space

+++ MatrixMath.cpp
@@ -16,8 +16,8 @@
 MatrixMath Matrix;			// Pre-instantiate
 
 // Matrix Printing Routine
-// Uses tabs to separate numbers under assumption printed float width won't cause problems
-void MatrixMath::Print(float* A, int m, int n, String label)
+// Uses tabs to separate numbers under assumption printed double width won't cause problems
+void MatrixMath::Print(double* A, int m, int n, String label)
 {
 	// A = input matrix (m x n)
 	int i, j;
@@ -34,7 +34,7 @@ void MatrixMath::Print(float* A, int m, int n, String label)
 	}
 }
 
-void MatrixMath::Copy(float* A, int n, int m, float* B)
+void MatrixMath::Copy(double* A, int n, int m, double* B)
 {
 	int i, j, k;
 	for (i = 0; i < m; i++)
@@ -46,7 +46,7 @@ void MatrixMath::Copy(float* A, int n, int m, float* B)
 
 //Matrix Multiplication Routine
 // C = A*B
-void MatrixMath::Multiply(float* A, float* B, int m, int p, int n, float* C)
+void MatrixMath::Multiply(double* A, double* B, int m, int p, int n, double* C)
 {
 	// A = input matrix (m x p)
 	// B = input matrix (p x n)
@@ -66,7 +66,7 @@ void MatrixMath::Multiply(float* A, float* B, int m, int p, int n, float* C)
 
 
 //Matrix Addition Routine
-void MatrixMath::Add(float* A, float* B, int m, int n, float* C)
+void MatrixMath::Add(double* A, double* B, int m, int n, double* C)
 {
 	// A = input matrix (m x n)
 	// B = input matrix (m x n)
@@ -81,7 +81,7 @@ void MatrixMath::Add(float* A, float* B, int m, int n, float* C)
 
 
 //Matrix Subtraction Routine
-void MatrixMath::Subtract(float* A, float* B, int m, int n, float* C)
+void MatrixMath::Subtract(double* A, double* B, int m, int n, double* C)
 {
 	// A = input matrix (m x n)
 	// B = input matrix (m x n)
@@ -96,7 +96,7 @@ void MatrixMath::Subtract(float* A, float* B, int m, int n, float* C)
 
 
 //Matrix Transpose Routine
-void MatrixMath::Transpose(float* A, int m, int n, float* C)
+void MatrixMath::Transpose(double* A, int m, int n, double* C)
 {
 	// A = input matrix (m x n)
 	// m = number of rows in A
@@ -108,14 +108,14 @@ void MatrixMath::Transpose(float* A, int m, int n, float* C)
 			C[m * j + i] = A[n * i + j];
 }
 
-void MatrixMath::Scale(float* A, int m, int n, float k)
+void MatrixMath::Scale(double* A, int m, int n, double k)
 {
 	for (int i = 0; i < m; i++)
 		for (int j = 0; j < n; j++)
 			A[n * i + j] = A[n * i + j] * k;
 }
 
-int MatrixMath::RPYtoMatrix(float* A, int m, int n, float R, float P, float Y)
+int MatrixMath::RPYtoMatrix(double* A, int m, int n, double R, double P, double Y)
 {//converts from RPY angles to a 4x4 matrix, m,n must =4
 	if(m==4 && n==4)
 	{
@@ -127,10 +127,10 @@ int MatrixMath::RPYtoMatrix(float* A, int m, int n, float R, float P, float Y)
 			}
 		}
 		//This is ugly but it will do for now
-		A[n*0+0]=cos(R)*cos(P); A[n*0+1]=cos(R)*sin(P)*sin(Y)-sin(R)*cos(Y); A[n*0+2]=cos(R)*sin(P)*cos(Y)+sin(R)*sin(Y); A[n*0+3]=(float)0;
-		A[n*1+0]=sin(R)*cos(P);  A[n*1+1]=sin(R)*sin(P)*sin(Y)-cos(R)*cos(Y); A[n*1+2]=sin(R)*sin(P)*cos(Y)-cos(R)*sin(Y);  A[n*1+3]=(float)0;
-		A[n*2+0]=-sin(P); 		A[n*2+1]=cos(P)*sin(Y);					   A[n*2+2]=cos(P)*cos(Y);				       	    A[n*2+3]=(float)0;
-		A[n*3+0]=(float)0;		A[n*3+1]=(float)0;	  					           A[n*3+2]=(float)0;				                    A[n*3+3]=(float)1;
+		A[n*0+0]=cos(R)*cos(P); A[n*0+1]=cos(R)*sin(P)*sin(Y)-sin(R)*cos(Y); A[n*0+2]=cos(R)*sin(P)*cos(Y)+sin(R)*sin(Y); A[n*0+3]=(double)0;
+		A[n*1+0]=sin(R)*cos(P);  A[n*1+1]=sin(R)*sin(P)*sin(Y)-cos(R)*cos(Y); A[n*1+2]=sin(R)*sin(P)*cos(Y)-cos(R)*sin(Y);  A[n*1+3]=(double)0;
+		A[n*2+0]=-sin(P); 		A[n*2+1]=cos(P)*sin(Y);					   A[n*2+2]=cos(P)*cos(Y);				       	    A[n*2+3]=(double)0;
+		A[n*3+0]=(double)0;		A[n*3+1]=(double)0;	  					           A[n*3+2]=(double)0;				                    A[n*3+3]=(double)1;
 		return 1;
 	}
 	else
@@ -145,14 +145,14 @@ int MatrixMath::RPYtoMatrix(float* A, int m, int n, float R, float P, float Y)
 //	 NUMERICAL RECIPES: The Art of Scientific Computing.
 // * The function returns 1 on success, 0 on failure.
 // * NOTE: The argument is ALSO the result matrix, meaning the input matrix is REPLACED
-int MatrixMath::Invert(float* A, int n)
+int MatrixMath::Invert(double* A, int n)
 {
 	// A = input matrix AND result matrix
 	// n = number of rows = number of columns in A (n x n)
 	int pivrow;		// keeps track of current pivot row
 	int k, i, j;		// k: overall index along diagonal; i: row index; j: col index
 	int pivrows[n]; // keeps track of rows swaps to undo at end
-	float tmp;		// used for finding max value and making column swaps
+	double tmp;		// used for finding max value and making column swaps
 
 	for (k = 0; k < n; k++)
 	{
